Fixed tests
The main problem seems to be that the ExponentialSmoothedAverage initialization in C#, was not being done accordingly exactly to the TA-Libs implementation. In order to achieve that, I had to change it so that EMA initializes with a SMA for the first n periods. Additionally I had to increase the warmup period to 120, instead of the theoretical 35, in order for it to work properly